ControlPanel prototyping
