+++ src/miniapps/ControlCentreMiniApp/BatteryControl.css
@@ -1,0 +1,17 @@
+.battery-control-container {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+  
+  .battery-icon {
+    width: 20px;
+    height: 14px;
+    object-fit: contain;
+  }
+  
+  .battery-percentage {
+    font-size: 12px;
+    font-weight: 500;
+  }
+
