+++ src/miniapps/ControlCentreMiniApp/DisplayControl.css
@@ -1,0 +1,53 @@
+.display-control-container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+  
+  .display-title {
+    font-size: 13px;
+    font-weight: 500;
+    margin-right: 8px;
+  }
+  
+  /* The brightness slider */
+  .display-slider {
+    flex: 1;
+    -webkit-appearance: none;
+    appearance: none;
+    height: 4px;
+    border-radius: 2px;
+    outline: none;
+    margin: 0;
+    cursor: pointer;
+    background: linear-gradient(to right, #999 0%, #fff 100%);
+  }
+  
+  .display-slider::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    width: 14px;
+    height: 14px;
+    background: #ccc;
+    border-radius: 50%;
+    cursor: pointer;
+    transition: background 0.2s ease;
+  }
+  
+  .display-slider::-webkit-slider-thumb:hover {
+    background: #aaa;
+  }
+  
+  /* Firefox */
+  .display-slider::-moz-range-thumb {
+    width: 14px;
+    height: 14px;
+    background: #ccc;
+    border-radius: 50%;
+    cursor: pointer;
+    transition: background 0.2s ease;
+  }
+  
+  .display-slider::-moz-range-thumb:hover {
+    background: #aaa;
+  }
+
