+++ src/miniapps/ControlCentreMiniApp/MusicControl.css
@@ -1,0 +1,149 @@
+.music-control-container {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  /* TOP SECTION: Album + Title + Artist */
+  .music-info-top {
+    display: flex;
+    align-items: center;
+    margin-bottom: 8px;
+  }
+  
+  .album-artwork {
+    margin-right: 8px;
+  }
+  
+  .album-artwork img {
+    width: 40px;
+    height: 40px;
+    border-radius: 4px;
+    object-fit: cover;
+  }
+  
+  .song-details {
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+  }
+  
+  .song-title-container {
+    overflow: hidden;
+    white-space: nowrap;
+    margin-top: -2px;
+  }
+  
+  .song-title-text {
+    font-size: 12px;
+    font-weight: bold;
+    display: inline-block;
+    white-space: nowrap;
+  }
+  
+  .song-artist {
+    font-size: 11px;
+    color: #555;
+    margin-top: 2px;
+    overflow: hidden;
+    white-space: nowrap;
+  }
+  
+  .song-artist-text {
+    display: inline-block;
+    white-space: nowrap;
+  }
+  
+  /* MIDDLE SECTION: Timeline & Times */
+  .progress-container {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    margin: 8px 0;
+  }
+  
+  .time-text {
+    font-size: 10px;
+    color: #444;
+    width: 34px;
+    text-align: center;
+  }
+  
+  .progress-bar {
+    flex: 1;
+    -webkit-appearance: none;
+    appearance: none;
+    height: 4px;
+    border-radius: 2px;
+    outline: none;
+    margin: 0;
+    cursor: pointer;
+    /* background is overridden inline by style={} */
+  }
+  
+  /* Chrome / Safari */
+  .progress-bar::-webkit-slider-runnable-track {
+    height: 4px;
+    border-radius: 2px;
+  }
+  
+  .progress-bar::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    width: 2px;
+    height: 16px;
+    background: #000;
+    margin-top: -6px;
+    border-radius: 1px;
+    cursor: pointer;
+    transition: transform 0.1s ease;
+  }
+  
+  .progress-bar::-webkit-slider-thumb:hover {
+    transform: scale(1.1);
+  }
+  
+  /* Firefox */
+  .progress-bar::-moz-range-track {
+    height: 4px;
+    border-radius: 2px;
+  }
+  
+  .progress-bar::-moz-range-thumb {
+    width: 2px;
+    height: 16px;
+    background: #000;
+    border: none;
+    border-radius: 1px;
+    cursor: pointer;
+    transition: transform 0.1s ease;
+    margin: 0;
+  }
+  
+  .progress-bar::-moz-range-thumb:hover {
+    transform: scale(1.1);
+  }
+  
+  /* BOTTOM SECTION: Playback buttons */
+  .music-buttons {
+    display: flex;
+    justify-content: center;
+    gap: 16px;
+    margin-top: 4px;
+  }
+  
+  .music-btn {
+    background: none;
+    border: none;
+    cursor: pointer;
+    padding: 4px;
+    transition: opacity 0.2s ease;
+  }
+  
+  .music-btn:hover {
+    opacity: 0.7;
+  }
+  
+  .music-btn img {
+    width: 16px;
+    height: 16px;
+  }
+
