+++ src/miniapps/ControlCentreMiniApp/QuickControls.css
@@ -1,0 +1,36 @@
+/* Container for the row of items */
+.quick-controls-container {
+    display: flex;
+    flex-wrap: wrap; /* if you want them to wrap on narrower containers */
+    gap: 12px;
+  }
+  
+  /* Each block: Wi-Fi / Bluetooth / AirDrop / Focus, etc. */
+  .quick-control-block {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  
+    background: rgba(255, 255, 255, 0.25);
+    border-radius: 10px;
+    padding: 6px 10px;
+    min-width: 80px;
+    cursor: pointer;
+    transition: background 0.2s ease;
+  }
+  
+  .quick-control-block:hover {
+    background: rgba(255, 255, 255, 0.4);
+  }
+  
+  .quick-control-title {
+    font-size: 14px;
+    font-weight: 500;
+  }
+  
+  .quick-control-sub {
+    font-size: 11px;
+    color: #666;
+    margin-top: 2px;
+  }
+
