+++ src/miniapps/ControlCentreMiniApp/SoundControl.css
@@ -1,0 +1,104 @@
+.sound-control-container-outer {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .sound-title {
+    font-size: 13px;
+    font-weight: 500;
+    margin-bottom: 8px;
+  }
+  
+  /* The row that holds volume slider, etc. */
+  .sound-control-container {
+    display: flex;
+    align-items: center;
+  }
+  
+  .volume-slider-container {
+    position: relative;
+    width: 100%;
+    height: 20px;
+  }
+  
+  /* The speaker icon pinned to the track */
+  .volume-icon-ontrack {
+    position: absolute;
+    opacity: 0.6;
+    left: 3px;
+    top: 50%;
+    transform: translateY(-50%);
+    z-index: 2;
+    width: 14px;
+    height: 14px;
+    user-select: none;
+  }
+  
+  .volume-slider {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 20px;
+    margin: 0;
+    cursor: pointer;
+    outline: none;
+    border-radius: 10px;
+    -webkit-appearance: none;
+    appearance: none;
+    --volume: 50%;
+    --white-stop: 50%;
+  }
+  
+  .volume-slider::-webkit-slider-runnable-track {
+    height: 20px;
+    border-radius: 10px;
+    border: 1px solid #bbb;
+    background: linear-gradient(
+      to right,
+      #fff 0%,
+      #fff var(--white-stop),
+      #ccc var(--white-stop),
+      #ccc 100%
+    );
+  }
+  
+  .volume-slider::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 19px;
+    height: 18px;
+    border-radius: 10px;
+    background: #fff;
+    border: 0.5px solid rgba(153, 153, 153, var(--thumb-border-opacity, 1));
+    box-shadow: 0 0 2px rgba(0, 0, 0, 0.3);
+    cursor: pointer;
+    transition: border-color 0.3s ease;
+  }
+  
+  /* Firefox track */
+  .volume-slider::-moz-range-track {
+    height: 20px;
+    border-radius: 10px;
+    border: 1px solid #bbb;
+    background: linear-gradient(
+      to right,
+      #fff 0%,
+      #fff var(--white-stop),
+      #ccc var(--white-stop),
+      #ccc 100%
+    );
+  }
+  
+  /* Firefox thumb */
+  .volume-slider::-moz-range-thumb {
+    width: 20px;
+    height: 20px;
+    border-radius: 10px;
+    background: #fff;
+    border: 1px solid rgba(153, 153, 153, var(--thumb-border-opacity, 1));
+    box-shadow: 0 0 2px rgba(0, 0, 0, 0.3);
+    cursor: pointer;
+    transition: border-color 0.3s ease;
+  }
+
